pyproject.toml
[project]
name = "openshelf"
version = "0.6.0"
description = "RESTful API for managing a library catalog. The system allows you to manage information about books, authors, readers, and book loans."
authors = [
    {name = "maxim.w",email = "max.wojw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "python-socketio (>=5.16.1,<6.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "fastapi-cache (>=0.1.0,<0.2.0)",
    "cachetools (>=7.0.3,<8.0.0)",
    "structlog (>=25.5.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
